Allow the EC2 driver to use public AMI's.
When using the amazon-ec2 gem method "describe_images", if
both :owner_id and :image_id are specified, it means that
it will only fetch information about images matching
:image_id that are owned by owner :owner_id.  Unfortunately,
this means public AMI's are excluded.  In the case where
we know the image_id, don't set the owner_id; the number of
responses can only be 0 (meaning that image_id wasn't found)
or 1 (meaning that image_id was found), so it shouldn't
pose a performance problem.  With this in place I was able
to start an EC2 instance from a public AMI.

Signed-off-by: Chris Lalancette <clalance@redhat.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@962310 13f79535-47bb-0310-9956-ffa450edef68