Generate correct URLs when not running at the root context.
This patch fixes the URLs generated by deltacloud-core when it
is proxied or deployed via Passenger or TorqueBox at a non-root
context.

It mainly consists of using url_for for all URLs and improving the
url_for logic to handle the cases where it gets applied multiple
times to the same url. This happens in a few places indirectly
because link_to also calls url_for on the passed-in URL.

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@982930 13f79535-47bb-0310-9956-ffa450edef68