Require rubygems before sinatra in config.ru
Sinatra is a gem so we need to require it after requiring
rubygems, not before. This happens to work when the server is
started via deltacloudd command because that command requires
rubygems.

When the server is started without the deltacloudd command (like
in TorqueBox or Passenger) the problem is exposed.

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@991370 13f79535-47bb-0310-9956-ffa450edef68