Canonicalize instance states.
For some drivers, we were passing the raw string we got from
the cloud provider through as the instance state.  However,
that doesn't fit in with trying to make the drivers act consistent,
so make sure all of the drivers have a translation layer between
the cloud provider's states and the deltcloud API states.  After
this series, there are only 3 valid deltacloud states: "RUNNING",
"STOPPED", and "PENDING".  We may want to think about adding
additional states in the future, but these cover most of the
current use-cases.

Signed-off-by: Chris Lalancette <clalance@redhat.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@999133 13f79535-47bb-0310-9956-ffa450edef68