Namespace dynamically generated classes by type.
This resolves an issue where based on the results of one request,
the client will create an Instance class as a subclass of StateFullObject
then, based on the results of another request try to create an Instance
as a subclass of ActionObject, causing a superclass mismatch error.

This patch groups the dynamic classes inside nested containing classes
inside DeltaCloud::API, preventing collisions.

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1041320 13f79535-47bb-0310-9956-ffa450edef68