Drivers now declare the capabilities they provide by implementing methods for each collection.
This patch makes the with_capability feature in rabbit actually work - before
this, all drivers would appear to provide most capabilities, since the check
was a simple respond_to?, and most capabilities were defined in the base_driver,
and overridden by the drivers. With this change, those empty definitions are
removed from the base_driver.

The next step is to expose the available capabilities with the entry_points.
I've started on that, but it introduces top level api changes, so I'll send
a separate message to the list for discussion.

Toby

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1044338 13f79535-47bb-0310-9956-ffa450edef68