Allow setting the provider on the server side, and implement region overriding in the ec2 driver based off of the provider (rev 2).
The provider can now be set on the commandline to deltacloudd via -P/--provider
(putting it in ENV['API_PROVIDER']), and also from the client via a header
(putting it in Thread.current[:provider]).

The ec2 driver is currently the only driver that understands providers, and
looks in Thread.current[:provider], then ENV['API_PROVIDER'], then falls back
to a default (currently 'us-east-1', the AWS default).

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1044351 13f79535-47bb-0310-9956-ffa450edef68