Fix long username for RHEVM driver
This patch is to fix a problem with Base64.b64encode for credentials in RHEVM driver, due to a wrong way in pack to add \n every 60 chars i had to use this hack.
The issue is not present in Ruby 1.9.x in fact they use a method called strict_encode64() that is the same as my hack.

Signed-off-by: Francesco Vollero <fvollero@redhat.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1132619 13f79535-47bb-0310-9956-ffa450edef68