Rationalize param declarations.
It seems pretty odd that param initialization can do
different things depending on the type of the 3rd argument.
I think it is easier to understand if the 3rd argument always
means 'options', and the 4th argument always means 'description'.

Luckily most of the callers were already conforming to this
pattern; just a few of them had to be changed to add an additional
argument between the [:required, :optional] declaration and
the description.

Along with this, rationalize the param declarations in features.rb
so that the options are set to [], instead of nil.  This matches
what param declarations do by default.

Signed-off-by: Chris Lalancette <clalance@redhat.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1132999 13f79535-47bb-0310-9956-ffa450edef68