Remove EC2 instance tagging and load balancer registering.
During an EC2 create_instance call, two optional things
might be done; tagging the instance with the name provided,
and registering the new instance with a loadbalancer.

Unfortunately, there is a big problem with this; the instance
ID does not always exist immediately after the ec2.launch_instances
call returns.  This seems to be a generic problem with EC2,
as I've run into it several times.

Because of this, we *cannot* tag the instance or register
the instance with the load balancer right within the create_instance
call, since we can't be sure the instance ID yet exists.
Luckily, the loss of either function shouldn't be that huge;
the name can never be retrieved (since there is no generic tag
support), and you can always register an instance with a load
balancer later.

Therefore, totally remove these operations from create.  Not
only does this make create_instance more reliable, it restores
the "one deltacloud API -> one cloud backend call" semantic
that we try to preserve.

v2: Remove more of the tagging code that is now unnecessary; also
remove the "load_balancer" feature from the ec2 driver

Signed-off-by: Chris Lalancette <clalance@redhat.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1136077 13f79535-47bb-0310-9956-ffa450edef68