Add condor-cloud driver to deltacloud.
This patch adds a condor-cloud driver to deltacloud.  This is a little
different from most drivers in that it directly drives condor and is
part of the control logic for the cloud itself.

This deals with most of the things David pointed out except for the
base_models.rb issue.  I also left the encoding of uuid and config
server because they need to go into the instance itself and really this
is the best place to do that.

    Ian

Signed-off-by: Michal fojtik <mfojtik@redhat.com>

Removed base_model, reogranized file layout of driver

Signed-off-by: Michal fojtik <mfojtik@redhat.com>

Added quite a few recent fixes.  Bridge is br0 as per docs, run submit
command as 'condor' user, renamed some stuff etc.

Signed-off-by: Ian Main <imain@redhat.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1151112 13f79535-47bb-0310-9956-ffa450edef68