mock driver: split into driver and client
This rewrite takes care of a number of issues with the driver (e.g., newly
created blobs would not show up in the bucket), and cleans up the code so
that it hopefully becomes clearer what driver writers need to do.

The driver now also uses the objects stored in the file system verbatim,
instead of modifying loaded objects on read.

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1160852 13f79535-47bb-0310-9956-ffa450edef68