Change the initscript to use the mock driver by default.
Since it is owned by the top-level deltacloud-core package
(and not the deltacloud-core-ec2 subpackage), there is no
guarantee that the ec2 driver will work.  Instead, use the
mock driver by default (which we know will work), and leave
it up to the user to install subpackages if they want to
use other drivers.

Signed-off-by: Chris Lalancette <clalance@redhat.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/deltacloud/trunk@1171525 13f79535-47bb-0310-9956-ffa450edef68