VSphere: The 'ToolsUnavailable' exception is now reported as backend error
Previously the ToolsUnavailable exception from VSPhere was
reported as error 500 (Internal Server Error), which isn't correct.
This exception is caused by lack of VMWare tools in machine. These
tools are required for the 'reboot' operation.
The exception is now reported as error 502 (Backend Error).
Appropriate error message is being passed to the client in XML