Core: Replaced 'check_capability' method with Sinatra::Base 'set' method
The 'check_capability' method set a class variable in Sinatra::Rabbit
to a lambda that was evaluated when server was starting.
However this class variable was not replaced when client switched the
driver using HTTP headers.

This patch will replace it with 'set' and now the lambda is evaluated
for every request in Sinatra::Rabbit (sinatra-rabbit 1.0.7)