Python client: simplify making requests
This removes the code duplication in the GET, POST, DELETE and PUT
methods and adds a generic `do_request` method that where the caller can
specify the HTTP method.

We will need that in order to respect the HTTP methods returned by the
Deltacloud API calls instead of hardcoding them.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>