EC2 tests: change the VCR matcher to ignore time- and user-dependent params
Requests to EC2 contain the parameters AWSAccessKeyId, Signature, and
Timestamp which are dependent on the user recording the tests and the time
of the recording. By ignoring these requests params the recordings become
independent of the access key/secret of the user making the recording, and
of how long ago the recordings were made.

That also obviates the need for Timecop-style time manipulation