CIMI: Initial implementation of 'attributes' in ResourceMetadata
Provider can define an extra attributes for each resource. In case
of Deltacloud API we do need provide Consumer a way how to define
the 'realm_id' when creating Machine.

This patch will add 'realm' attribute that hold information about
the 'realm' (eg. where Machine is located).

This patch also add 'machineImage' attribute to Machine that point
client to MachineImage that was used for creating Machine