Core: Add root_type attr to Image and HardwareProfile
The new 'root_type' attribute in case of Image contains information
about the root device type. There are two possible values:

persistent - Volume or other 'persistent' storage type (EBS)
transient  - Default device that is part of the Image itself (instance-store)

In case of HardwareProfile this attribute is presented as a fixed
'property' called 'root_type'. The possible values are the same as above.

The hardware profile and image root_type attributes *must* match in order
to create a new instance. In other case an HTTP 400 (Bad Request) will be
returned to client.

API-Change: Added the 'root_type' attribute to Image and HardwareProfile