Related to DMTF cimi mantis 1974:
The state of a volume being created in fgcp is called "DEPLOYING", the
Deltacloud API documentation seems to leave it to the provider, with
examples of what ec2 returns. Ec2 (and cimi) define "CREATING". Aligning
what the fgcp driver returns with that. Same for snapshot "PENDING",
volume "NORMAL" and "BACKING_UP" states.
Cimi intended we use their state names for same states. Deltacloud has
"IN-USE" for volumes that are "AVAILABLE". Aligning this state with CIMI
for all providers. Whether a volume has been attached can be retrieved
through other attributes/apis.
Also, for fgcp, generate volume name if empty string specified (happens
with GUI), and align with cimi CREATING/AVAILABLE for snapshot states.