CIMI: Initial inclusion of CIMI action models
The way of how we currently create CIMI entities is not
really nice. We parse the raw JSON/XML and then we are
trying to create the CIMI entity.

Besides the code looks ugly, there is no way to add validation
for required attributes without make it even more uglier.

Then in MachineTemplate model, we can leverage the '.from_xml' and
the '.from_json' methods to create the action model and then use it
to create the CIMI entity (MachineTemplate in this case)