CIMI: Fixed undefined 'default_frontend' error in CIMI tests
In case when db/provider model is accessed outside the Deltacloud
server context (CIMI tests, CIMI client, etc...) the
Deltacloud.default_frontend method is not set, because 'deltacloud_rack'
is not required.

This patch will add 'fallback' mode where if you use this model outside
server context, we try to get the default driver in same way as we
get the provider value.