* server/tests/test_helper.rb: save/load scrubbed config data for each driver
Anonymize user/password and give tests a chance to modify other things
before writing.

For some tests, like RHEV-M, the provider URL contains sensitive
information, like the hostname, which should not be recorded. At the same
time, we need to preserve the URL structure so that VCR matches requests on
playback.